==> 02_Seoul_Crime/01_data_outline.ipynb <==
# 서울시 범죄 현황 데이터 개요 확인

import pandas as pd
import numpy as np

#### 🔰 데이터 읽기

- `thousands` param은 단위 구분 문자(콤마)가 찍힌 숫자를 문자로 인식할 수 있기 때문에 설정으로 넣어준다.

	- 숫자값들이 콤마(,)를 사용하고 있어서 문자로 인식될 수 있다.
	- 천단위 구분(thousands=',')이라고 알려주면 컴마를 제거하고 숫자형으로 읽어온다.

crime_raw_data = pd.read_csv("../data/02. crime_in_Seoul.csv", thousands=',', encoding='euc-kr')
crime_raw_data.head()

#### 🔰 기본 정보 확인

- `info()` 메서드로 데이터의 개요를 확인한다.

crime_raw_data.info()

👆 RangeIndex: 65534 entries, 0 to 65533 </br>

- 자료는 310개인데 인덱스가 6만개가 넘는다.

- 특정 컬럼에서 unique 조사

crime_raw_data["죄종"].unique()

👆 "죄종" 컬럼에 nan 값이 들어가 있음을 확인할 수 있다.

- `isnull()` 메서드를 이용해 Value가 Null인지 아닌지 True, False로 결과를 출력해 본다.

crime_raw_data["죄종"].isnull()

- crime_raw_data로 마스킹을 씌워 True인 값들만 DataFrame형태로 출력한다.

crime_raw_data[crime_raw_data["죄종"].isnull()].head()

👆 Index [310]부터 NaN값으로 채워진 것을 확인할 수 있다.

- 이럴 때는 NaN을 제거하는 것이 아니라 NaN값을 제외하고 필요한 데이터만 가져온다.

crime_raw_data[crime_raw_data["죄종"].notnull()]

crime_raw_data = crime_raw_data[crime_raw_data["죄종"].notnull()]
crime_raw_data.info()

👆 NaN을 정리하고 나니 데이터의 크기가 2MB에서 12.1KB가 되었다.

- head()와 tail()로 값을 찍어 문제가 없는지 확인한다.

crime_raw_data.head()

crime_raw_data.tail()

- 이 데이터의 형태로는 의미있는 insight를 도출하기 어렵다.
- 형태가 세로축에 서울기 구이름이 있고, 가로축에 5대 범죄 수치가 있으면 좋을 것같다.
- 즉, 단순 나열된 데이터를 원하는 데이터 열로 재배치할 필요가 있다.

- 이것을 쉽게 해주는 것이 Pandas의 `pivot_table`이다.

crime_raw_data.to_csv("./result_data/01_crime_raw_data.csv", encoding='utf-8', index=False)
