==> 02_Seoul_Crime/02_data_pivot_table.ipynb <==
# 서울시 범죄 현황 데이터 정리

> pivot_table을 사용해 원하는 DataFrame 형태를 만든다.

import pandas as pd

crime_raw_data = pd.read_csv("./result_data/01_crime_raw_data.csv")
crime_raw_data.head()

#### 🔰 Index 설정

- 경찰서 이름인 "구분" 컬럼을 Index로 설정해 정리한다.

- 집계 함수의 Default가 '평균(mean)'이므로 사건의 합이 기록되도록 `aggfunc` param에 sum을 사용한다.

import numpy as np

crime_station = crime_raw_data.pivot_table(
	index="구분",
	columns=['죄종', '발생검거'],
	aggfunc=np.sum
)

crime_station.head()

❓ sum 함수를 왜 리스트에 넣었을까??? 2개 이상의 함수를 적용하는 것도 아니고 단일 함수인데...

crime_station = crime_raw_data.pivot_table(
	index="구분",
	columns=['죄종', '발생검거'],
	aggfunc=[np.sum]
)

crime_station.head()

👆 문제는 컬럼이 Multi로 표현된다는 것이다!

- Multi Columns Index
	- 맨 위 "건수"에서 부터 level[0] 컬럼, 그 아래로 level[1], level[2], level[3] 컬럼까지 4개의 컬럼이 존재

crime_station.columns

👆 컬럼을 조회해 보면 'sum'과 '건수'가 10개씩 반복되는 것을 알 수 있다.

- Multi Index에 대한 접근 방법

crime_station['sum', '건수', '강도', '검거']

crime_station['sum', '건수', '강도', '검거'][:5]

#### 🔰 다중 컬럼에서 특정 컬럼 제거

- 필요 없는 'sum'과 '건수' 컬럼을 날린다.

crime_station.columns = crime_station.columns.droplevel([0, 1])
crime_station.columns

crime_station.head()

crime_station.index

👆 현재 Index는 경찰서의 이름으로 되어 있다.

- 우리의 목적은 "구별" 범죄 현황이므로 경찰서 이름을 구 이름으로 변경해야 한다.

- 경찰서의 이름으로 구 이름을 알아내기 위해 Googlemaps를 활용한다.
