==> 02_Seoul_Crime/03_data_alteration.ipynb <==
# Google Maps를 이용한 데이터 변경 및 정리

import pandas as pd

crime_raw_data = pd.read_csv("./result_data/01_crime_raw_data.csv")
crime_raw_data.head()

#### 🔰 DataFrame 재편성

> raw_data를 pivot_table()을 사용해 DataFrame을 재편성하고 "건수"를 합으로 재집계한다.

import numpy as np

crime_station = crime_raw_data.pivot_table(
	index="구분",
	columns=["죄종", "발생검거"],
	aggfunc=np.sum
)

crime_station.head()

crime_station.columns = crime_station.columns.droplevel([0])
crime_station.head()

#### 🔰 컬럼 추가

> 구글맵을 이용해 경찰서 이름이 소속된 구 이름과 위도, 경도 정보를 저장할 컬럼을 생성한다.

- 새로 추가한 컬럼의 values를 모두 NaN으로 채운다.

crime_station["구별"] = np.nan
crime_station["lat"] = np.nan # latitude, 위도
crime_station["lng"] = np.nan # longitude, 경도
crime_station.head()

#### 🔰 Python의 반복문

> Pandas에 잘 맞춰진 반복문용 함수 `iterrows()`에 대해 알아본다.

- pandas DataFrame은 index(row)와 column, 2차원 형태이다.

	이럴 때 for문을 사용하면, n번째라는 지정을 반복해야 하므로 가독률이 떨어진다.

- pandas DataFrame으로 반복문을 만들때 `itterows()` 함수를 사용하면 편리하다.

	받을 때, 인덱스와 행의 내용으로 나누어 받는 것만 주의할 것!

for idx, row in crime_station.iterrows():
    print(idx, end='\n\n')
    print(row)
    print('-'*30)

count = 0
for idx, row in crime_station.iterrows():
    print(count, idx, type(idx))
    print("서울" + idx + "경찰서")
    print('-'*30)
    count += 1

#### 🔰 새로운 데이터 추가

> 경찰서 이름으로 googlemaps에서 위치 정보를 얻어와 필요한 데이터로 새로 추가한 컬럼 values를 채운다.

import googlemaps

gmaps = googlemaps.Client(key="")

crime_station.index

cnt = 0
for idx, row in crime_station.iterrows():
    if idx == "동작":
        station_name = idx + "경찰서"
    else:
        station_name = "서울" + idx + "경찰서"
    
    temp = gmaps.geocode(station_name, language="ko")
    
    # 주소 정보 얻어오기
    tmp_gu = temp[0].get("formatted_address")
    
    # 위치(위도, 경도) 가져오기
    lat = temp[0].get("geometry")["location"]["lat"]
    lng = temp[0].get("geometry")["location"]["lng"]
    
    # 각 컬럼에 값 채우기
    crime_station.loc[idx, "구별"] = tmp_gu.split()[2]
    crime_station.loc[idx, "lat"] = lat
    crime_station.loc[idx, "lng"] = lng
    
    print(cnt)
    cnt += 1

crime_station["구별"]

crime_station.head()

crime_station.tail()

#### 🔰 멀티 컬럼 합치기

> "죄종"과 "발생검거" 두 줄의 컬럼을 하나로 합친다.

- 다음과 같은 결과를 기대한다. -> "강도검거", 강도발생", ..

- `df.columns.get_level_values(__level)` : 컬럼 레벨별로 값을 가져온다.

crime_station.columns.get_level_values(0)

crime_station.columns.get_level_values(1)

len(crime_station.columns.get_level_values(0)),	len(crime_station.columns.get_level_values(1))

crime_station.columns.get_level_values(0)[2] + crime_station.columns.get_level_values(1)[2]

tmp = [
	crime_station.columns.get_level_values(0)[n] + crime_station.columns.get_level_values(1)[n]
	for n in range(len(crime_station.columns.get_level_values(0)))
]
tmp

crime_station.columns = tmp
crime_station.head()

crime_station.to_csv("./result_data/03_crime_in_seoul_raw_data.csv", encoding='utf-8')
