==> 02_Seoul_Crime/04_Gu_data_arrange.ipynb <==
# 구별 범죄 현황 데이터 생성

import pandas as pd

crime_anal_station = pd.read_csv(
    "./result_data/03_crime_in_seoul_raw_data.csv", encoding='utf-8'
    , index_col=0 # "구분" 컬럼을 Index로 설정
)
crime_anal_station.head()

#### 🔰 구별 데이터로 정리

> 서울은 한 구에 경찰서가 두 곳인 구가 있으므로 구의 이름으로 다시 정렬한다.

- 서울시 구는 25개, 서울 내 경찰서는 31곳

crime_anal_station.info()

import numpy as np

crime_anal_gu = pd.pivot_table(crime_anal_station, index="구별", aggfunc=np.sum)
crime_anal_gu.head()

crime_anal_gu.info()

- 필요 없는 컬럼 제거

del crime_anal_gu["lat"]
del crime_anal_gu["lng"]

crime_anal_gu.head()

#### 🔰 검거율 데이터 생성

- 하나의 컬럼을 다른 컬럼으로 나누기

crime_anal_gu["강도검거"] / crime_anal_gu["강도발생"]

- 다수의 컬럼을 다른 컬럼으로 나누기

crime_anal_gu[["강도검거", "살인검거"]].div(crime_anal_gu["강도발생"], axis=0).head()

- 다수의 컬럼을 다수의 컬럼으로 각각 나누기

num = ["강간검거", "강도검거", "살인검거", "절도검거", "폭력검거"]
den = ["강간발생", "강도발생", "살인발생", "절도발생", "폭력발생"]

crime_anal_gu[num].div(crime_anal_gu[den].values).head()

- "검거율" 컬럼 생성

target = ["강간검거율", "강도검거율", "살인검거율", "절도검거율", "폭력검거율"]

crime_anal_gu[target] = crime_anal_gu[num].div(crime_anal_gu[den].values) * 100
crime_anal_gu.head()

- 필요 없는 컬럼 제거

crime_anal_gu.drop(["강간검거", "강도검거", "살인검거", "절도검거", "폭력검거"], axis=1, inplace=True)
crime_anal_gu.head()

👆 검거율 값 중 100이 넘는 것들이 있다.

- 검거율 100 이상의 수치는 강제로 100을 만들어 준다.

crime_anal_gu[crime_anal_gu[target] > 100]

# 100보다 큰 숫자 찾아서 바꾸기

crime_anal_gu[crime_anal_gu[target] > 100] = 100
crime_anal_gu.head()

#### 🔰 컬럼명 변경

crime_anal_gu.rename(
	columns={
		"강간발생": "강간",
		"강도발생": "강도",		
		"살인발생": "살인",		
		"절도발생": "절도",		
		"폭력발생": "폭력",		
	},
	inplace=True
)
crime_anal_gu.head()

crime_anal_gu.to_csv("./result_data/04_crime_anal_gu.csv", encoding='utf-8')
