==> 02_Seoul_Crime/05_data_normalization.ipynb <==
# 범죄 데이터 정규화

> 범죄별 발생 건수를 0과 1 사이 값으로 정규화한다.

import pandas as pd

crime_anal_gu = pd.read_csv("./result_data/04_crime_anal_gu.csv", index_col=0)
crime_anal_gu.head()

#### 🔰 구별 데이터에서 발생건수 정규화 데이터 생성

- 범죄의 경중에 따라 발생 건수의 차이가 크다.
- 살인은 한 자리 수 발생, 절도나 폭력은 네 자리 수 발생

- 발생 건수를 0과 1 사이로 정규화하여 최대값이 1이 되면 상호 비교가 용이해진다.

- 정규화 : 최고값은 1, 최소값은 0

crime_anal_gu["강도"] / crime_anal_gu["강도"].max()

col1 = ["살인", "강도", "강간", "절도", "폭력"]

crime_anal_norm = crime_anal_gu[col1] / crime_anal_gu[col1].max()
crime_anal_norm.head()

- 새로 만든 DataFrame에 검거율 컬럼도 추가

col2 = ["강간검거율", "강도검거율", "살인검거율", "절도검거율", "폭력검거율"]

crime_anal_norm[col2] = crime_anal_gu[col2]
crime_anal_norm.head()

#### 🔰 서울시 CCTV 데이터 추가

> 서울시 CCTV 자료에서 구별 인구수와 CCTV수를 가져와 추가한다.

pd.read_csv("../01_Seoul_CCTV/result_data/04_merge_data_result.csv").head(1)

result_CCTV = pd.read_csv(
    "../01_Seoul_CCTV/result_data/04_merge_data_result.csv",
    index_col="구별", encoding="utf-8"
)
result_CCTV.head()


crime_anal_norm[["인구수", "CCTV"]] = result_CCTV[["인구수", "소계"]]
crime_anal_norm.head()

#### 🔰 지표 데이터 추가

- np.mean()

	- axis=0 열(culumn), axis=1 행(row)을 기준으로 평균을 계산

import numpy as np

arr1D = np.array([0.357143, 1.000000, 1.000000, 0.977118, 0.733773])
arr1D

np.mean(arr1D)

arr2D = np.array([
    [0.357143, 1.000000, 1.000000, 0.977118, 0.733773],
    [0.285714, 0.358974, 0.310078, 0.477799, 0.463880]
])
arr2D

np.mean(arr2D, axis=0), np.mean(arr2D, axis=1)

> 정규화된 범죄발생 건수 전체의 평균을 구해서 '범죄'의 대표값으로 사용한다.

- "범죄" 컬럼을 만들고 5대 범죄 발생건수의 평균값을 할당한다.

col = ["살인", "강도", "강간", "절도", "폭력"]

crime_anal_norm["범죄"] = np.mean(crime_anal_norm[col], axis=1) # axis=1 행을 따라서 연산하는 옵션 
crime_anal_norm.head()

> 5대 범죄 검거율의 평균을 구해 '검거'의 대표값으로 사용한다.

- "검거" 컬럼을 만들고 5대 범죄 검거율의 평균값을 할당한다.

col = ["강간검거율", "강도검거율", "살인검거율", "절도검거율", "폭력검거율"]

crime_anal_norm["검거"] = np.mean(crime_anal_norm[col], axis=1)
crime_anal_norm

crime_anal_norm.to_csv("./result_data/05_crime_anal_norm.csv", encoding='utf-8')
